user_app/rescue_connect/ml_backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
httpx>=0.27.0
pillow>=11.0.0
supabase>=2.10.0
python-dotenv>=1.0.1
openai>=1.55.0
google-genai>=1.0.0
imagehash>=4.3.0
resend>=2.0.0

# New dependencies for integrated modules
spacy>=3.7.0
rapidfuzz>=3.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
ultralytics>=8.0.0
geopy>=2.4.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
